Updated style.css and index.html
index.html:
Moved aria-current to the correct location.

style.css:
Changed the nav bar to be at the very top and be sticky.
Put nav elements into the center of the nav bar.
Made the nav bar take up the whole width of the screen
Made the navbar scroll if it does not all fit on one page
Hide overflow on the indival items of the navbar
Removed the margin and padding on html and body and made overflow automatic
Highted the section of the navbar that the user is currently on